+++ packages/authentication/modules/access/useLogout/index.ts
@@ -1,28 +1,28 @@
 import {
-  ACCESS_COOKIE_NAME,
+  ACCESS_KEY_NAME,
   LOGOUT_EVENT,
-  REFRESH_COOKIE_NAME,
+  REFRESH_KEY_NAME,
   eventEmitter,
+  removeTokenAsync,
 } from '@baseapp-frontend/utils'
 
 import { useQueryClient } from '@tanstack/react-query'
-import Cookies from 'js-cookie'
 
 import { MFA_API_KEY } from '../../../services/mfa'
 import { USER_API_KEY } from '../../../services/user'
-import { UseLogoutOptions } from './types'
+import type { UseLogoutOptions } from './types'
 
 const useLogout = ({
-  cookieName = ACCESS_COOKIE_NAME,
-  refreshCookieName = REFRESH_COOKIE_NAME,
+  accessKeyName = ACCESS_KEY_NAME,
+  refreshKeyName = REFRESH_KEY_NAME,
   onLogout,
   emitLogoutEvent = true,
 }: UseLogoutOptions = {}) => {
   const queryClient = useQueryClient()
 
-  const logout = () => {
-    Cookies.remove(cookieName)
-    Cookies.remove(refreshCookieName)
+  const logout = async () => {
+    await removeTokenAsync(accessKeyName)
+    await removeTokenAsync(refreshKeyName)
     queryClient.resetQueries({ queryKey: USER_API_KEY.getUser() })
     queryClient.resetQueries({ queryKey: MFA_API_KEY.default })
     onLogout?.()