+++ packages/core/jest.config.js
@@ -1,1 +1,0 @@
-module.exports = require('@baseapp-frontend/config/jest.config.js')